build: rearrange dependencies of buildSrc dsl module
