+++ buildSrc/src/main/kotlin/dependencies/TrackAnalysisDependencies.kt
@@ -1,0 +1,9 @@
+package dependencies
+
+object FirebaseDependencies {
+    const val boomVersion = "32.7.2"
+    const val boom = "com.google.firebase:firebase-bom"
+
+    const val crashlyticsVersion = "18.6.2"
+    const val crashlytics = "com.google.firebase:firebase-crashlytics:$crashlyticsVersion"
+}
